support float as routers per node

+++ src/run.py
@@ -1,7 +1,7 @@
 import os
 from Simulate import Simulate
 
-average_routers_per_node = int(
+average_routers_per_node = float(
     os.environ.get('average_routers_per_node', "Not set"))
 routers_tier_2 = int(os.environ.get('routers_tier_2', "Not set"))
